pyproject.toml
[tool.poetry]
name = "bear-exporter"
version = "0.1.0"
description = "Export Bear notes to Markdown"
authors = ["Cade Ekblad-Frank <cade@e-f.me>"]

[tool.poetry.scripts]
bear-exporter = 'bear_exporter.cli:main'

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
python-slugify = "^4.0.1"
tablib = {extras = ["cli"], version = "^2.0.0"}
click-option-group = "^0.5.2"
peewee = "^3.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-datadir = "^1.3.1"
faker = "^4.1.3"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
